last-update(uint64) should be transferred as string in JSON issue #328

+++ dots_server/db_models/data/vendor_mapping.go
@@ -8,7 +8,7 @@ type VendorMapping struct {
 	DataClientId int64     `xorm:"'data_client_id' not null"`
 	VendorId     int       `xorm:"'vendor_id' not null"`
 	VendorName   string    `xorm:"'vendor_name'"`
-	LastUpdated  int       `xorm:"'last_updated' not null"`
+	LastUpdated  uint64    `xorm:"'last_updated' not null"`
 	Created      time.Time `xorm:"created"`
 	Updated      time.Time `xorm:"updated"`
 }
